pyproject.toml
[project]
name = "llms-generator"
version = "0.6.9"
description = "Generate llms.txt files from websites using AI extraction"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "defusedxml>=0.7.0",
]

[project.scripts]
llms-generator = "llms_generator.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llms_generator"]

[tool.ruff]
line-length = 100
target-version = "py310"
